Added Testing for form validation, added attributes to index.js to prevent warnings
Added testing for the booking form to validate it works.Added test ids to the input to be able to find them efficiently in test.js. Attributes in for the BrowserRouter in the injdex.js is to stop warning about future use of the code and thus to future proof the work.