Form Update using Formik and Yup
Wanted to use a library for the form and yup to ensure that validation and proper input was done. i chose to use Formik and Yup for its simplicity to integrate and its compatibility. Removed some test cases and some of the attributes are no longer needed. Added CSS to accommodate for the new error outputs.